docs() #210 Improve dependency installation description
Add notes on how to install the correct software versions in Ubuntu
18.04.

The pip install command gives errors regarding failing of building
wheels on Ubuntu 18.04 when using Python 3.6. This is caused by the
`wheel` shell command not being installed by default. However when using
the correct Python version 3.8 this does not happen (also not on Ubuntu
20.04) so I did not include a mention of this in the document.